{"schema":"decodex/commit/1","summary":"Split runtime MCP and dashboard asset modules","authority":"manual"}
