+++ apps/decodex/src/orchestrator/operator_http/assets/html.rs
@@ -1,708 +1,28 @@
+mod markup;
+mod scripts;
+mod styles;
+
 use std::sync::LazyLock;
+
+use self::{
+	markup::{BODY, HEAD, TAIL},
+	scripts::SCRIPT_PARTS,
+	styles::STYLE_PARTS,
+};
 
 pub(in crate::orchestrator::operator_http) static OPERATOR_DASHBOARD_HTML: LazyLock<String> =
 	LazyLock::new(|| {
-		[
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/head.html"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/tokens.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/tokens/theme.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/tokens/base.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/masthead.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/masthead/shell.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/masthead/metrics.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/masthead/toolbar.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/notices.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/notices/trigger.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/notices/panel.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/controls.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/controls/theme.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/controls/transport.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/controls/links.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/flow.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/flow/strip.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/foundation/flow/steps.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/workspace.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/workspace/frame.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/workspace/panels.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/workspace/cards.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/cards.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/cards/header.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/cards/meta.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/projects.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/projects/table.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/projects/rows.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/projects/content.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/layout/controls.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/controls/visibility.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/controls.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/meta.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/pool/activity.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/pool/summary.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/pool.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/profile.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/table.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/table/sorting.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-surface.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-surface/base.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-surface/state.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-surface/animations.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-identity.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-identity/name-controls.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-identity/profile-controls.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-identity/plan.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-usage.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-usage/windows.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-usage/state-credit.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-status.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-status/status.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/accounts/row-status/meta-slot.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/activity.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/activity/child-layout.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/activity/child-metrics.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/activity/status-tone-grid.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/details.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/details/fields.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/details/disclosure.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/details/phases.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/details/fold-panels.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/details/empty-state.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/responsive.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/responsive/wide.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/responsive/medium.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/responsive/mobile.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/styles/responsive/motion.css"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/body.html"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/boot.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/time.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/time/timestamps.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/time/stream.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/time/duration.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/run-state.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/run-state/telemetry.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/run-state/status.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/formatting/queue.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/preferences.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/preferences/layout.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/preferences/accounts.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/preferences/projects.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/preferences/controls.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/labels.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/labels/text.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/labels/theme.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/labels/fields.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/labels/status.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/details.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/details/facts.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/details/state.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/details/animation.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/dom.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/history.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/history/identity.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/history/lanes.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/history/run-summaries.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/issue-identity.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/freshness.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/metrics.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/lane-meta.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/current-lane.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/current-lane/metrics.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/current-lane/phases.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/render-primitives/lifecycle/current-lane/facts.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/selection.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/selection/account.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/selection/confirmation.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/core.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/random-name.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/display.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/capacity.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/token.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/privacy.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/identity/numbers.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/usage/records.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/usage/pool.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/usage/time.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/usage/profile.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/activity.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/expansion.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/window-state.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/status.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/credits.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/meta.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/window-render.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/profile/toggle.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/rows.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/sort-controls.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/mode.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/summary.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/render.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/run-meta.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/sorting.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/accounts/pool/render-account-pool.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/history.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/child-agent/context.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/child-agent/buckets.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/child-agent/lifecycle.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/child-agent.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/child-agent/summary.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/notices.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/notices/readiness.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/notices/list.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/notices/accounts.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/notices/warnings.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/activity/notices/render.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/core/derived-state.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/core/header.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/projects/model.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/projects/table-controls.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/projects/sorting.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/projects/render.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/flow.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/overview/programs.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/cards.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/current.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/recent.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/worktrees.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/worktrees/role.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/worktrees/fields.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/lanes/worktrees/render.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/clock.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/live-activity.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/socket.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/render.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/startup.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/events/accounts.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/events/projects.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/app/stream/events/document.js"
-			)),
-			include_str!(concat!(
-				env!("CARGO_MANIFEST_DIR"),
-				"/src/orchestrator/operator_dashboard/tail.html"
-			)),
-		]
-		.concat()
+		let mut html = String::new();
+
+		html.push_str(HEAD);
+		for style in STYLE_PARTS {
+			html.push_str(style);
+		}
+		html.push_str(BODY);
+		for script in SCRIPT_PARTS {
+			html.push_str(script);
+		}
+		html.push_str(TAIL);
+
+		html
 	});
